Refactor team board reports into a single report namespace
